// src/api/orval/model/vpnServerGroupsDtoVpnServerGroupDto.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */

export interface VpnServerGroupsDtoVpnServerGroupDto {
  id?: number;
  /** @nullable */
  name?: string | null;
  sortOrder?: number;
  /** @nullable */
  serverIds?: number[] | null;
}